Add of routers and bug fix
-fix of bug that would not log in users correctly and display the wrong
data
-add of router for pricing endpoint
-add of router for games endpoint
