+++ games/index.js
@@ -1,0 +1,2 @@
+const {router} = require('./router');
+module.exports = {router};
